requirements.txt
asgiref==3.8.1
Django==5.1.5
sqlparse==0.5.3
tzdata==2025.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
python-decouple==3.8
django-rest-knox==4.2.0
django-filter==23.5
